edit lesson 8 file

+++ Notes/lesson_8.java
@@ -11,18 +11,17 @@ public class lesson_8 {
     //! Add positive numbers, To stop enter negative number
     System.out.println("Enter some positive numbers to add them together or enter negative number to stop the program: ");
     double sum =0.0; 
-    float value;
+    double value;
     boolean nonNegative = true;
     while(nonNegative){
-      value = input.nextFloat();
+      value = input.nextDouble();
       if (value < 0){
         nonNegative=false;
       } else {
         sum += value;
       }
     }
-    System.out.printf("Sum of input numbers is %.2f",sum);
-
+    System.out.printf("Sum of input numbers is %f",sum);
 
   }
 }
